requirements.txt
aniso8601==10.0.1
APScheduler==3.10.4
attrs==25.3.0
bcrypt==4.0.1
black==23.7.0
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorlog==6.7.0
coverage==7.9.2
cryptography==41.0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
flake8==6.0.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
flask-restx==1.1.0
gunicorn==21.2.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.2.6
openpyxl==3.1.2
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
psutil==5.9.5
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
pyscard==2.0.7
pyserial==3.5
pytest==7.4.2
pytest-cov==4.1.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.36.2
requests==2.31.0
rpds-py==0.26.0
schedule==1.2.0
six==1.17.0
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.0.4
Werkzeug==3.1.3
XlsxWriter==3.1.9
pyusb==1.2.1
